move blurb mutations to Blurbs
